requirements.txt
aiomqtt>=2.0.0
asyncpg>=0.29.0
PyYAML>=6.0
